Update 1.modules.md
On this  doc url: https://nuxt.com/docs/guide/directory-structure/modules

the last `p tag` contains the text `The local modules are registered by alphabetical order. You can change the order by prefixing with a number if front of each directory:` which has a typo at the end of the sentence. ` a number [ if ] front of each`. the `if` should be `in`.